+++ modules/base_bt_nodes.py
@@ -1,4 +1,5 @@
 from enum import Enum
+import random
 # BT Node List
 class BTNodeList:
     CONTROL_NODES = [        
@@ -225,3 +226,111 @@ class AssignTask(SyncAction):
             return Status.FAILURE        
         else:                        
             return Status.SUCCESS    
+
+
+class _IsTaskCompleted(SyncAction):
+    def __init__(self, name, agent):
+        super().__init__(name, self._update)
+
+    def _update(self, agent, blackboard, task_id_key = 'task_id'):        
+        _task_id = blackboard.get(task_id_key)
+        if _task_id is None:
+            raise ValueError(f"[{self.name}] Error: No {task_id_key} found in the blackboard!")
+        
+        task = agent.tasks_info[_task_id]
+        if task.completed is True:
+            blackboard[task_id_key] = None
+            # agent.set_assigned_task_id(None)
+            return Status.SUCCESS  
+        return Status.FAILURE          
+    
+class _IsArrivedAtTask(SyncAction):
+    def __init__(self, name, agent):
+        super().__init__(name, self._update)
+
+    def _update(self, agent, blackboard, task_id_key = 'task_id', arrive_threshold = 10.0):
+        _task_id = blackboard.get(task_id_key)
+        if _task_id is None:
+            raise ValueError(f"[{self.name}] Error: No {task_id_key} found in the blackboard!")
+
+        agent_position = agent.position
+        task_position = agent.tasks_info[_task_id].position
+        # Calculate norm2 distance
+        distance = (task_position - agent_position).length()
+
+        if distance < agent.tasks_info[_task_id].radius + arrive_threshold: # Agent reached the task position                                                
+            return Status.SUCCESS  
+        return Status.FAILURE      
+    
+class _MoveToTask(SyncAction):
+    def __init__(self, name, agent):
+        super().__init__(name, self._update)
+
+    def _update(self, agent, blackboard, task_id_key = 'task_id'):
+        _task_id = blackboard.get(task_id_key)
+        if _task_id is None:
+            raise ValueError(f"[{self.name}] Error: No {task_id_key} found in the blackboard!")
+        
+        # Move towards the task position
+        task_position = agent.tasks_info[_task_id].position        
+        agent.follow(task_position)  
+
+        return Status.RUNNING     
+    
+class _ExecuteTask(SyncAction):
+    def __init__(self, name, agent):
+        super().__init__(name, self._update)
+
+    def _update(self, agent, blackboard, task_id_key = 'task_id'):
+        _task_id = blackboard.get(task_id_key)
+        if _task_id is None:
+            raise ValueError(f"[{self.name}] Error: No {_task_id} found in the blackboard!")
+        
+        # Task Execution
+        agent.tasks_info[_task_id].reduce_amount(agent.work_rate)
+        agent.update_task_amount_done(agent.work_rate) 
+
+        return Status.RUNNING     
+        
+class _ExecuteTaskWhileFollowing(SyncAction):
+    def __init__(self, name, agent):
+        super().__init__(name, self._update)
+
+    def _update(self, agent, blackboard, task_id_key = 'task_id'):
+        _task_id = blackboard.get(task_id_key)
+        if _task_id is None:
+            raise ValueError(f"[{self.name}] Error: No {_task_id} found in the blackboard!")
+        
+        # Task Execution
+        agent.tasks_info[_task_id].reduce_amount(agent.work_rate)
+        agent.update_task_amount_done(agent.work_rate) 
+
+        # Move towards the task position (gradually)
+        task_position = agent.tasks_info[_task_id].position
+        agent.follow(task_position)  
+
+        return Status.RUNNING     
+    
+class _ExploreArea(SyncAction):
+    def __init__(self, name, agent):
+        super().__init__(name, self._update)
+        self.random_move_time = float('inf')
+        self.random_waypoint = (0, 0)
+
+    def _update(self, agent, blackboard, agent_max_random_movement_duration = 1000, exploration_area = {'x_min': 0, 'x_max': 1400, 'y_min': 0, 'y_max': 1000}, sampling_time = 1.0):
+        # Move towards a random position
+        if self.random_move_time > agent_max_random_movement_duration:
+            self.random_waypoint = self.get_random_position(exploration_area['x_min'], exploration_area['x_max'], exploration_area['y_min'], exploration_area['y_max'])
+            self.random_move_time = 0 # Initialisation
+        
+        self.random_move_time += sampling_time   
+        agent.follow(self.random_waypoint)         
+        return Status.RUNNING
+        
+    def get_random_position(self, x_min, x_max, y_min, y_max):
+        pos = (random.randint(x_min, x_max),
+                random.randint(y_min, y_max))
+        return pos
+    
+    def halt(self):
+        self.random_move_time = float('inf')